pyproject.toml
[project]
name = "sneaky-vaccine-tg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.20.0.post0",
    "motor>=3.7.1",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]
